requirements.txt
Django>=5.0,<6
youtube-transcript-api>=0.6.2
google-generativeai>=0.8.0
yt-dlp>=2024.1.0
djangorestframework>=3.14
requests>=2.31
python-dotenv>=1.0
chromadb>=0.5
sentence-transformers>=2.2
pypdf>=4.0
python-docx>=1.0
tqdm>=4.0
